Crate agora

A library for managing signal processing networks.

The main concept behind this library is that individual processing steps will be implemented as Nodes which are arranged into a Graph. Nodes can receive input signals, both internal and external, and are expected to produce an output signal.

Once you create and register a Node with a Graph instance, you can connect any two Nodes in an arbitrary manner. You can configure any Node to receive external input here as well. The connections and set of Nodes can be modified at any time, and changes will take effect the next time you process a buffer.

Assumptions and caveats

This library is built to work with double-precision floating point signals.

Nodes can only consume one stream of input data and produce one stream of output data.

Signal graphs must be acyclic - no feedback loops are permitted.

Buffer length and sample rate must be constant (per Graph and Node instance) at compile time. The tradeoff here is between safety (via the type system) versus flexibility, and for this library safety is preferred.

Resource usage

The time complexity of processing a buffer of data is linear in the number of nodes, and memory usage is linear in both the number of nodes and the specified buffer size.

The time complexity of adding a node or removing a connection is constant (amortized).

The time complexity of replacing or removing a node is linear in the number of nodes.

The time complexity of connecting two nodes is proportional to the number of nodes times the number of existing connections. This is due to the recalculation of the critical path through the graph.

Example

struct DcOffsetNode { offset: f64 }
impl Node<BUF_LEN, 44_100> for DcOffsetNode {
    fn process(&mut self, buffer: &mut Buffer<BUF_LEN>) {
        for sample in buffer { *sample += self.offset }
    }
}

let mut graph = Graph::default();
let node_0 = graph.add_node(Box::new(DcOffsetNode { offset: 1.0 }));

let mut buffer = [0.0; BUF_LEN];
graph.process(&mut buffer);
assert_eq!(buffer, [1.0; BUF_LEN]);

let node_1 = graph.add_node(Box::new(DcOffsetNode { offset: 2.0 }));
graph.connect(node_0, node_1);

buffer = [0.0; BUF_LEN];
graph.process(&mut buffer);
assert_eq!(buffer, [3.0; BUF_LEN]);

Afterthoughts

There are a few design decisions made in this library that may deserve more consideration.

Use of const generics for buffer size and sample rate

While this is very nice in terms of safety (incompatible processing elements are found immediately at compile time), it is not flexible at all to different platforms and environments. There are many differences between different host operating systems in terms of handling audio, and it may be preferable to support some level of configurability if building an application intended to be cross-platform.

Use of standard library

The Rust standard library provides an excellent developer experience - however, there are some platforms which do not support it, or at least the memory allocation mechanisms it uses. Building a no_std version of this library would not be possible in its current incarnation, and would require significant changes to the code.

I/O limitations

Nodes are only given one buffer that is shared between input and output. Most platforms support (at least) stereo I/O, and many operations (such as dynamically mixing signals) would be made much easier with more flexibility here. The cpal library (used here in the example application) utilizes interleaved audio - maybe this would be the right way to handle it.

Parallelization

This implementation is single-threaded. It would not require significant refactoring to parallelize, but some work would still be necessary.

Panics

The only (anticipated) panics in this library are in Graph::process, essentially as assertions on the validity of the graph’s construction. These could be replaced with returned errors without much refactoring effort, but since they are checking the internal logic of the library it may be best to keep them as is.

Structs

Graph

Orchestrate a group of processing nodes and use that network to handle signal data.

NodeReference

Provides a reference to a Node added to a Graph.

Enums

Error

Describes the type of error encountered when making a connection between Nodes.

Constants

EXTERNAL_INPUT

Use this value in the Graph::connect method to indicate that a node receives external input.

Traits

Node

Performs an isolated processing step in a signal graph.

Type Definitions

Buffer

A buffer of signal samples.